+++ modules/module-postgres-storage/test/src/storage.test.ts
@@ -99,7 +99,9 @@ for (let storageVersion of TEST_STORAGE_VERSIONS) {
 
       const options: storage.BucketDataBatchOptions = {};
 
-      const batch1 = await test_utils.fromAsync(bucketStorage.getBucketDataBatch(checkpoint, [globalBucket], options));
+      const batch1 = await test_utils.fromAsync(
+        bucketStorage.getBucketDataBatch(test_utils.testCheckpoint(checkpoint), [globalBucket], options)
+      );
       expect(test_utils.getBatchData(batch1)).toEqual([
         { op_id: '1', op: 'PUT', object_id: 'test1', checksum: 2871785649 }
       ]);
@@ -111,7 +113,7 @@ for (let storageVersion of TEST_STORAGE_VERSIONS) {
 
       const batch2 = await test_utils.fromAsync(
         bucketStorage.getBucketDataBatch(
-          checkpoint,
+          test_utils.testCheckpoint(checkpoint),
           [{ ...globalBucket, start: BigInt(batch1[0].chunkData.next_after) }],
           options
         )
@@ -127,7 +129,7 @@ for (let storageVersion of TEST_STORAGE_VERSIONS) {
 
       const batch3 = await test_utils.fromAsync(
         bucketStorage.getBucketDataBatch(
-          checkpoint,
+          test_utils.testCheckpoint(checkpoint),
           [{ ...globalBucket, start: BigInt(batch2[0].chunkData.next_after) }],
           options
         )
@@ -143,7 +145,7 @@ for (let storageVersion of TEST_STORAGE_VERSIONS) {
 
       const batch4 = await test_utils.fromAsync(
         bucketStorage.getBucketDataBatch(
-          checkpoint,
+          test_utils.testCheckpoint(checkpoint),
           [{ ...globalBucket, start: BigInt(batch3[0].chunkData.next_after) }],
           options
         )