+++ modules/module-postgres/test/src/wal_stream_utils.ts
@@ -228,7 +228,7 @@ export class WalStreamTestContext implements AsyncDisposable {
       start = BigInt(start);
     }
     const syncConfigContent = this.getSyncConfigContent();
-    const { checkpoint } = await this.storage!.getCheckpoint();
+    const checkpoint = await this.storage!.getCheckpoint();
     const map = [bucketRequest(syncConfigContent, bucket, start)];
     const batch = this.storage!.getBucketDataBatch(checkpoint, map);
     const batches = await test_utils.fromAsync(batch);